6　環境佈署、前端呈現/requirements.txt
absl-py==0.8.1
alabaster==0.7.12
argh==0.26.2
asn1crypto==1.2.0
astor==0.8.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.7.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.0
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cloudpickle==1.2.2
colorama==0.4.1
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.5
dlib==19.8.1
docutils==0.15.2
entrypoints==0.3
flake8==3.7.9
future==0.18.2
gast==0.2.2
google-auth==1.7.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
imagesize==1.1.0
importlib-metadata==1.1.0
imutils==0.5.3
intervaltree==3.0.2
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.14.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==19.2.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
line-bot-sdk==1.8.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.4
numpydoc==0.9.1
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
packaging==19.2
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.1
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.1
psutil==5.6.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pymongo==3.9.0
Pympler==0.7
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.5
pyrsistent==0.15.6
PySocks==1.7.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.2
python-language-server==0.31.2
pytz==2019.3
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
QDarkStyle==2.7
QtAwesome==0.6.0
qtconsole==4.6.0
QtPy==1.9.0
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.0
Rtree==0.8.3
scipy==1.3.3
Send2Trash==1.5.0
six==1.13.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
Sphinx==2.2.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==4.0.0
spyder-kernels==1.8.1
sqlparse==0.3.0
tensorboard==1.9.0
tensorflow==1.9.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
ujson==1.35
urllib3==1.25.7
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
yapf==0.28.0
zipp==0.6.0
